fix: proxy interface should be public
