+++ utam-core/src/test/java/utam/core/framework/base/BasePageObjectTests.java
@@ -9,12 +9,8 @@ package utam.core.framework.base;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import static utam.core.framework.base.BasicElementBuilder.getUnwrappedElement;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 import org.testng.annotations.Test;
 import utam.core.MockUtilities;
 import utam.core.element.Actionable;
@@ -28,17 +24,6 @@ import utam.core.selenium.element.LocatorBy;
  * @since 236
  */
 public class BasePageObjectTests {
-
-  @Test
-  public void testProxyElementCanBeUsedInsteadRoot() {
-    MockUtilities mockUtilities = new MockUtilities();
-    WebElement mockElement = mock(WebElement.class);
-    when(mockUtilities.getWebDriverMock().findElement(By.cssSelector("root")))
-        .thenReturn(mockElement);
-    TestProxyPageObject pageObject = mockUtilities.getFactory().create(TestProxyPageObject.class);
-    UnionType element = pageObject.getRoot();
-    element.isPresent();
-  }
 
   @Test
   public void testGetProxyDoesNotThrow() {
@@ -69,29 +54,18 @@ public class BasePageObjectTests {
 
   interface UnionType extends Actionable, Clickable { }
 
-  @SuppressWarnings("WeakerAccess")
-  @PageMarker.Find(css = "root")
-  public static class TestProxyPageObject extends BaseRootPageObject {
-
-    public UnionType getRoot() {
-      return getProxy(getRootElement(), UnionType.class);
-    }
-  }
-
   @SuppressWarnings("unused")
   static class ElementDeclaration extends BasePageObject {
+
+    @ElementMarker.Find(css = "css")
+    ElementLocation one;
+    @ElementMarker.Find(css = "css", nullable = true)
+    ElementLocation two;
+    @ElementMarker.Find(css = "css", expand = true)
+    ElementLocation three;
 
     ElementDeclaration(MockUtilities utilities, Locator root) {
       utilities.getFactory().bootstrap(this, utilities.getElementAdapter(), root);
     }
-
-    @ElementMarker.Find(css = "css")
-    ElementLocation one;
-
-    @ElementMarker.Find(css = "css", nullable = true)
-    ElementLocation two;
-
-    @ElementMarker.Find(css = "css", expand = true)
-    ElementLocation three;
   }
 }
